+++ webview-ui/src/components/account/AccountWelcomeView.tsx
@@ -20,7 +20,7 @@ export const AccountWelcomeView = () => {
 			</p>
 
 			<VSCodeButton className="w-full mb-4" disabled={isLoginLoading} onClick={handleSignIn}>
-				Sign up with MitraHelix
+				Sign up with MitraH
 				{isLoginLoading && (
 					<span className="ml-1 animate-spin">
 						<span className="codicon codicon-refresh" />