+++ webview-ui/src/components/common/TelemetryBanner.tsx
@@ -21,10 +21,10 @@ export const TelemetryBanner: React.FC = () => {
 
 	return (
 		<div className="bg-banner-background text-banner-foreground px-3 py-2 flex flex-col gap-1 shrink-0 mb-1 relative text-sm m-4">
-			<h3 className="m-0">Help Improve MitraHelix</h3>
+			<h3 className="m-0">Help Improve Mitra Helix AI</h3>
 			<i>(and access experimental features)</i>
 			<p className="m-0">
-				MitraHelix collects error and usage data to help us fix bugs and improve the extension. No code, prompts, or
+				Mitra Helix AI collects error and usage data to help us fix bugs and improve the extension. No code, prompts, or
 				personal information is ever sent.
 			</p>
 			<p className="m-0">