+++ webview-ui/src/components/mcp/configuration/tabs/installed/ConfigureServersView.tsx
@@ -24,12 +24,11 @@ const ConfigureServersView = () => {
 					Model Context Protocol
 				</VSCodeLink>{" "}
 				enables communication with locally running MCP servers that provide additional tools and resources to extend
-				MitraHelix's capabilities. You can use{" "}
+				MitraH's capabilities. You can use{" "}
 				<VSCodeLink href="https://github.com/modelcontextprotocol/servers" style={{ display: "inline" }}>
 					community-made servers
 				</VSCodeLink>{" "}
-				or ask MitraHelix to create new tools specific to your workflow (e.g., "add a tool that gets the latest npm
-				docs").{" "}
+				or ask MitraH to create new tools specific to your workflow (e.g., "add a tool that gets the latest npm docs").{" "}
 				<VSCodeLink href="https://x.com/sdrzn/status/1867271665086074969" style={{ display: "inline" }}>
 					See a demo here.
 				</VSCodeLink>